src/requirements.txt
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
filelock @ file:///croot/filelock_1700591183607/work
fsspec==2024.5.0
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
greenlet==3.0.3
grpcio==1.60.0
grpcio-tools==1.60.0
idna @ file:///croot/idna_1714398848350/work
Jinja2 @ file:///croot/jinja2_1706733616596/work
joblib==1.3.2
MarkupSafe @ file:///croot/markupsafe_1704205993651/work
minepy==1.2.6
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mpmath @ file:///croot/mpmath_1690848262763/work
msgpack==1.0.8
networkx @ file:///croot/networkx_1690561992265/work
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
pandas==2.2.2
pillow @ file:///croot/pillow_1714398848491/work
protobuf==4.25.2
pynvim==0.4.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytz==2024.1
PyYAML @ file:///croot/pyyaml_1698096049011/work
requests @ file:///croot/requests_1707355572290/work
scikit-learn==1.4.0
scipy==1.12.0
six==1.16.0
sympy @ file:///croot/sympy_1701397643339/work
threadpoolctl==3.2.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
triton==2.3.0
typing_extensions @ file:///croot/typing_extensions_1715268824938/work
tzdata==2023.4
urllib3 @ file:///croot/urllib3_1715635851070/work
